feat(about): refactor About component to use FindMeOn for social links and remove unused styles
