+++ packages/web-qwik/src/components/main/about.module.scss
@@ -4,27 +4,3 @@
   width: 100%;
   min-height: calc(100vh - 3.5rem);
 }
-
-.link-container {
-  display: flex;
-  margin-block: 2rem;
-  justify-content: space-around;
-}
-
-.link {
-  all: unset;
-  box-sizing: content-box;
-  padding: 0.75rem;
-  border-radius: 0.25rem;
-  display: inline-flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 0.5rem;
-  cursor: pointer;
-  transition: opacity 200ms;
-
-  &:hover {
-    opacity: 0.8;
-    background-color: rgba(#59a7b5, 0.1);
-  }
-}
