+++ packages/web-qwik/src/components/main/find-me-on.module.scss
@@ -1,0 +1,23 @@
+.link-container {
+  display: flex;
+  margin-block: 2rem;
+  justify-content: space-around;
+}
+
+.link {
+  all: unset;
+  box-sizing: content-box;
+  padding: 0.75rem;
+  border-radius: 0.25rem;
+  display: inline-flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 0.5rem;
+  cursor: pointer;
+  transition: opacity 200ms;
+
+  &:hover {
+    opacity: 0.8;
+    background-color: rgba(#59a7b5, 0.1);
+  }
+}
